libsnapshot: Add support for Xor ops in Cow Format
This adds support for Xor Ops in the Cow Format. These represent store
possibly compressed data which must be xor'ed against the given section
of source data to be interpreted as a block in the new image. The cow
reader and writer do not have access to this data, so they assume the
provider and user of the data will handle the xor-ing.

Bug: 177104308
Test: cow_api_test (ReadWriteXor)
Change-Id: I7a313d2643773d6b81a878a3e5bc87e3bdfc387b

Former-commit-id: d83b2efb12ec13d0cc4faaa6c793985f17beac2b
Former-commit-id: 0b1f1d6485aceffa08c0129be5e18b32ea25e7bc