Avoid thread cache in unwinder.
The code in the fallback path calls pthread_key_create when using the
normal thread cache. However, this code is executed out of the linker,
which means that the call doesn't see keys created by the libc version
of pthread_key_create. As of now, simply avoid using the thread cache
to avoid this problem.

Bug: 189803009

Test: debuggerd -b on a media process on a 32 bit Android Go device
Test: and observe no crash.
Test: debuggerd unit tests pass.
Change-Id: I9ca1a55e44d3bb69d49450826d7d64d7a64145c3

Former-commit-id: 79643eb603d6587597495b603e929679e57b67a7 [formerly 49e5a765440604eac1208c831c1cd8c0f97cb09e]
Former-commit-id: 798245054441e5efffa00ec44268345b9fe53cbc