Revert "Re-land removing RECOVERY_AS_BOOT check for init_first_stage"
This reverts commit 73a259bb93cfd014a88432a90d53fc21439bf5f4 [formerly 988ffbaeb0279c69089edc15cc5f339c851689d4] [formerly 4818f74888bfe528d7a2090d18478c95aa88c497].

The previous commit might introduce a build break on legacy
devices that with a smaller /boot partition size, because it
outputs the first-stage init and the second-stage init binaries
in the following locations, respectively (when
BOARD_USES_RECOVERY_AS_BOOT is true).

    * $OUT/recovery/root/first_stage_ramdisk/init
    * $OUT/recovery/root/system/bin/init

The first one is not needed because there is already a symlink
under $OUT/recovery/root:  init -> /system/bin/init.

Bug: 193565271
Test: `lunch aosp_flame-userdebug`, `make bootimage_debug` then
      checks the output under $OUT/debug_ramdisk/.
Test: Performs the same check for `lunch aosp_bramble-userdebug`.
Change-Id: I6277a6c4c0980bd976cc35db6ca99eafc6bf3f5a

Former-commit-id: ab079af28f43c35977bdf4aa6d8328f4686bb6a9
Former-commit-id: a31f0748c3b471b246cf527e68dff3443dd2c1e0