[fs_mgr] Use shared libutils for remount
Bug: http://b/197965342

remount depends on libbinder.so which already depends on libutils.so.
When linking remount with libutils.a, the linker tries to fetch objects
from libutils.a to redefine undefined symbols in libbinder.so.  This
causes linker failure with upstream LLD.  Switching to shared
libutils.so instead doesn't increase the dependency closure for remount.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: I858042a08babe2c2db1fc8c7eeb5de62dda5f328

Former-commit-id: cdf919d1f0a6aec03f4aebcf0c0b9b52781492f2
Former-commit-id: 74017174bf141768cbe3177e1f4237b12138e9c2