Depend on KeyMint HAL via default
This allows for easier bumping of the KeyMint version level.

At the moment this change should have no effect: the same dependency
is used, just reached via a default rather than explicitly.

However, when the KeyMint version increases in the near future, using
this default should mean that no change is needed here: the default
definition will change to -V2 and this will be referenced here.

Test: TreeHugger
Change-Id: Ic250e5b91ee2b48cd7a05783ce21af16ae330ed1

Former-commit-id: c5b7d179c2c74f3479b60b6c2bedbc2be94f7c22
Former-commit-id: 1bcb6d8bd34ad5f92da14563e5cb7e71168b86d5